+++ src/models/command/lampSetTimeBasedCommand.ts
@@ -20,7 +20,7 @@ export class LampSetTimeBasedCommand extends BaseCommand {
    * @param time - The time to use for calculation of desired state
    * @param reason - You can provide a reason for clarity
    * @param disableAutomatic - If provided, the device will remain in the desired state for the given disable action.
-   * If unset the default value will be used: {@link SettingsService.settings.blockAutomaticHandlerDefaults}
+   * If undefined the default value will be used in case it's a non automatic action: {@link SettingsService.settings.blockAutomaticHandlerDefaults}
    */
   public constructor(
     source: CommandSource | BaseCommand,