+++ src/models/command/ledSetLightCommand.ts
@@ -17,7 +17,7 @@ export class LedSetLightCommand extends DimmerSetLightCommand {
    * @param on - The desired state-value
    * @param reason - You can provide a reason for clarity
    * @param disableAutomatic - If provided, the device will remain in the desired state for the given disable action.
-   * If unset the default value will be used: {@link SettingsService.settings.blockAutomaticHandlerDefaults}
+   * If undefined the default value will be used in case it's a non automatic action: {@link SettingsService.settings.blockAutomaticHandlerDefaults}
    * @param brightness - The desired brightness
    * @param transitionTime - The transition time during turnOn/turnOff
    * @param color - The desired color in 6 digit hex Code